+++ oxe-web-admin/src/component/item/Form.jsx
@@ -4,6 +4,7 @@ import Popup from "reactjs-popup";
 import { NotificationManager as nm } from "react-notifications";
 import { postRequest } from "../../utils/request.jsx";
 import Tab from "../tab/Tab.jsx";
+import DialogConfirmation from "../dialog/DialogConfirmation.jsx";
 import FormGlobal from "./form/FormGlobal.jsx";
 import FormQuestions from "./form/FormQuestions.jsx";
 import FormAnswers from "./form/FormAnswers.jsx";
@@ -14,8 +15,6 @@ import Item from "./Item.jsx";
 export default class Form extends Item {
 	constructor(props) {
 		super(props);
-
-		this.confirmDeletion = this.confirmDeletion.bind(this);
 
 		this.state = {
 			selectedMenu: null,
@@ -52,14 +51,13 @@ export default class Form extends Item {
 
 	confirmDeletion(close) {
 		const params = {
-			category: this.props.name,
+			id: this.props.form.id,
 		};
 
 		postRequest.call(this, "form/delete_form", params, () => {
-			document.elementFromPoint(100, 0).click();
 			nm.info("The form has been deleted");
 			close();
-			if (typeof this.props.afterDeletion !== "undefined") this.props.afterDeletion();
+			if (this.props.afterDeletion) this.props.afterDeletion();
 		}, (response) => {
 			nm.warning(response.statusText);
 		}, (error) => {
@@ -95,6 +93,16 @@ export default class Form extends Item {
 
 					<div className="col-md-3">
 						<div className={"right-buttons"}>
+							<DialogConfirmation
+								text={"Are you sure you want to delete this form?"}
+								trigger={
+									<button
+										className={"red-background"}>
+										<i className="fas fa-trash-alt"/>
+									</button>
+								}
+								afterConfirmation={() => this.confirmDeletion(close)}
+							/>
 							<button
 								className={"grey-background"}
 								data-hover="Close"